dom/tests/mochitest/pointerlock/file_exitMouseLockOnLoseFocus.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Bug 633602 - file_exitMouseLockOnLoseFocus.html</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
  </script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js">
  </script>
  <script type="application/javascript" src="mouselock_util.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="start();">
  <a target="_blank"
    href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
    Mozilla Bug 633602
  </a>
  <div id="div"></div>
  <pre id="test">
    <script type="text/javascript">
      /*
       * Test for Bug 633602
       * Tests if mouse is unlocked when the window losese focus
       * The mouse is locked to an element, if the window loses focus,
       * because a new tab was opened, or because the user agent lost focus,
       * the mouse should unlock itself.
       * For this test, in the callback passed to lock() a new window is open
       * to blur the focus. The locked element listens for the mozpointerlocklost
       * event, and tests if the mouse was unlocked when the window lost focus
       */

      SimpleTest.waitForExplicitFinish();
      var pointer = navigator.mozPointer;
      var pointerLockLost = false;
      var div, newWindow;

      document.addEventListener("mozfullscreenchange", function() {
        if (document.mozFullScreen && document.mozFullScreenElement === div) {
          // Mouse should be locked if the element is in fullscreen mode
          pointer.lock(div, function () {
            // Test if mouse is locked when success callback is dispatched
            ok(pointer.isLocked, "Success Callback - mouse should be locked");
            // Open a new window to switch focus and blur the current window
            // On opening a new window, the mozpointerlocklost should trigger
            newWindow = window.open();
          });
        } else {
          ok(pointerLockLost, "Mouselocklost event should trigger " +
            "when window loses focus and mouse should be unlocked");

          if (newWindow) {
            newWindow.close();
          }

          SimpleTest.finish();
        }
      }, false);

      function start() {
        SimpleTest.waitForFocus(function() {
          div = document.getElementById("div");

          div.addEventListener("mozpointerlocklost", function() {
            pointerLockLost = true;
          }, false);

          div.mozRequestFullScreen();
        });
      }
    </script>
  </pre>
</body>
</html>
